Bring AMDGPU memory model related bits in sync with phabricator reviews
Summary of changes:
  - Keyword for synchronization scope changed: synchscope -> syncscope
  - Syncscope argument size changed: 32 bits -> 8 bits
  - Minor improvements

Phabricator reviews:
  - https://reviews.llvm.org/D21723
  - https://reviews.llvm.org/D24577
  - https://reviews.llvm.org/D24623

+++ llvm/include/llvm/Bitcode/LLVMBitCodes.h
@@ -370,7 +370,7 @@ enum AtomicOrderingCodes {
 };
 
 /// Encoded SynchronizationScope values.
-enum AtomicSynchScopeCodes : unsigned {
+enum AtomicSynchScopeCodes : uint8_t {
   /// Encoded value for SingleThread synchronization scope.
   SYNCHSCOPE_SINGLETHREAD = 0,
